+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 ### Bugfixes:
 * Fix add-game prompt failing to appear if installed fresh.
 * Fix catwalk end segment being placed at the top of upwards-facing stairs.
+* #2048: Fix ceiling Faith Plates not functioning correctly.
 * #2040: Warn if no compiler files were found when exporting.
 * #1343: Clone fizzler items if connected to multiple standing/reclined fizzlers.